requirements.txt
python-dotenv
pandatic